Transfer AllBytes in caching + refactor the dropbox implementation
Modify the signature of the class to follow the base class
